ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
torch>=2.1.0
sentence-transformers>=2.6.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.7
langchain-milvus>=0.1.0
langgraph>=0.1.0
huggingface_hub>=0.23.0
pymilvus>=2.4.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
transformers>=4.40.0
accelerate>=0.29.0
peft>=0.10.0
